+++ code/events/TTTEvent.Settings.cs
@@ -1,0 +1,14 @@
+namespace TTTReborn.Events
+{
+    public static partial class TTTEvent
+    {
+        public static class Settings
+        {
+            /// <summary>
+            /// Occurs when server settings are changed.
+            /// <para>No data is passed to this event.</para>
+            /// </summary>
+            public const string Change = "tttreborn.settings.change";
+        }
+    }
+}
