+++ code/events/TTTEvent.Shop.cs
@@ -1,0 +1,14 @@
+namespace TTTReborn.Events
+{
+    public static partial class TTTEvent
+    {
+        public static class Shop
+        {
+            /// <summary>
+            /// Occurs when the shop is changed.
+            /// <para>No data is passed to this event.</para>
+            /// </summary>
+            public const string Change = "tttreborn.shop.change";
+        }
+    }
+}
