+++ code/player/TTTPlayer.Damage.cs
@@ -2,6 +2,7 @@ using System;
 
 using Sandbox;
 
+using TTTReborn.Events;
 using TTTReborn.Globals;
 using TTTReborn.Items;
