+++ code/player/TTTPlayer.Shop.cs
@@ -4,6 +4,7 @@ using System.Text.Json;
 
 using Sandbox;
 
+using TTTReborn.Events;
 using TTTReborn.Globals;
 using TTTReborn.Items;
 using TTTReborn.UI;
