+++ code/player/TTTPlayer.VoiceChat.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 
 using Sandbox;
 
-using TTTReborn.Globals;
+using TTTReborn.Events;
 using TTTReborn.Teams;
 
 namespace TTTReborn.Player
@@ -131,7 +131,7 @@ namespace TTTReborn.Player
             ConsoleSystem.Run((toggle ? "+" : "-") + "iv_voice");
         }
 
-        [Event(TTTEvent.Player.Role.OnSelect)]
+        [Event(TTTEvent.Player.Role.Select)]
         private static void OnSelectRole(TTTPlayer player)
         {
             if (!Host.IsServer)
