+++ code/player/TTTPlayer.cs
@@ -1,5 +1,6 @@
 using Sandbox;
 
+using TTTReborn.Events;
 using TTTReborn.Globals;
 using TTTReborn.Player.Camera;
 using TTTReborn.Roles;
