+++ code/ui/alivehud/quickshop/QuickShop.cs
@@ -4,7 +4,7 @@ using Sandbox;
 using Sandbox.UI;
 using Sandbox.UI.Construct;
 
-using TTTReborn.Globals;
+using TTTReborn.Events;
 using TTTReborn.Items;
 using TTTReborn.Player;
