+++ code/ui/generalhud/scoreboard/Scoreboard.cs
@@ -5,7 +5,7 @@ using Sandbox;
 using Sandbox.UI;
 using Sandbox.UI.Construct;
 
-using TTTReborn.Globals;
+using TTTReborn.Events;
 using TTTReborn.Player;
 
 namespace TTTReborn.UI
